+++ packages/webamp/js/types.ts
@@ -75,9 +75,27 @@ export type Band =
 
 export type Slider = Band | "preamp";
 
+export type CursorImage =
+  | {
+      type: "cur";
+      url: string;
+    }
+  | {
+      type: "ani";
+      frames: {
+        url: string;
+        rate: number;
+      }[];
+    };
+
 // TODO: Use a type to ensure these keys mirror the CURSORS constant in
 // skinParser.js
-export type Cursors = { [cursor: string]: string };
+export type Cursors = { [cursor: string]: CursorImage };
+
+export type AniFrame = {
+  url: string;
+  rate: number;
+};
 
 export type GenLetterWidths = { [letter: string]: number };